% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_mod}
\alias{demo_mod}
\title{Loads demo module included with the package}
\usage{
demo_mod()
}
\value{
Returns the demo module tracker object of class \code{openmpt}
}
\description{
Reads the file \code{cyberrid.mod}. It is a \href{https://en.wikipedia.org/wiki/Protracker}{ProTracker}
file create by \href{https://modarchive.org/index.php?request=view_profile&query=69138}{Jester}.
It is redistributed under the
\href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{Attribution Non-commercial Share Alike license}.
The music was part of an Amiga computer demo named 'Extension' and was originally released in 1993.
}
\examples{
mod <- demo_mod()
}
