% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_pattern_row_channel}
\alias{format_pattern_row_channel}
\alias{format_pattern_row_channel_command}
\alias{highlight_pattern_row_channel}
\alias{highlight_pattern_row_channel_command}
\title{Format a cell in a pattern table as text}
\usage{
format_pattern_row_channel(mod, pattern, row, channel, width, pad, ...)

format_pattern_row_channel_command(mod, pattern, row, channel, command, ...)

highlight_pattern_row_channel(mod, pattern, row, channel, width, pad, ...)

highlight_pattern_row_channel_command(mod, pattern, row, channel, command, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{pattern}{TODO}

\item{row}{TODO}

\item{channel}{TODO}

\item{width}{TODO}

\item{pad}{TODO}

\item{...}{Ignored}

\item{command}{One of \code{"note"}, \code{"instrument"}, \code{"volumeffect"}, \code{"effect"}, \code{"volume"}, or
\code{"parameter"}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
mod <- demo_mod()
format_pattern_row_channel(mod, 0L, 1L, 2L, 0L, TRUE)
format_pattern_row_channel_command(mod, 0L, 1L, 2L, "parameter")
highlight_pattern_row_channel(mod, 0L, 1L, 2L, 0L, TRUE)
highlight_pattern_row_channel_command(mod, 0L, 1L, 2L, "note")
# TODO
}
\seealso{
\code{\link[=pattern]{pattern()}}
}
