% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_duration_seconds}
\alias{get_duration_seconds}
\title{Get ModPlug Tracker module duration}
\usage{
get_duration_seconds(mod, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
