% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_metadata}
\alias{get_metadata}
\title{Get ModPlug Tracker module meta data}
\usage{
get_metadata(mod, key = "title", ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{key}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
