% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_order_pattern}
\alias{get_order_pattern}
\title{Get the pattern index of an \code{openmpt} module at a specific order index}
\usage{
get_order_pattern(mod, order, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{order}{Index of the position in the pattern sequence table (starts at \code{0}).}

\item{...}{Ignored}
}
\value{
Returns the \code{integer} index (starting at \code{0}) of the pattern at the
indicated \code{order} position.
}
\description{
A module contains a sequence table describing the order in which to play
patterns. This function returns the index of the patter at specific position
in the sequence table.
}
\examples{
mod <- demo_mod()
get_order_pattern(mod, 3L)
}
