% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portaudio.R
\name{has_audio_device}
\alias{has_audio_device}
\title{Test if there is an audio device}
\usage{
has_audio_device()
}
\value{
Returns a logical value.
}
\description{
Tests if an audio device is present on the system.
}
