% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modarchive.R
\name{modarchive_info}
\alias{modarchive_info}
\title{TODO}
\usage{
modarchive_info(mod_id, api = Sys.getenv("MODARCHIVE_API"))
}
\description{
TODO
}
\examples{
## These examples are wrapped in `tryCatch`
## as they attempt to access online resources

tryCatch({
  mod_info <- modarchive_info(41529)
  if (!is.null(mod_info)) mod_info$modarchive$module$url[[1]]
  elekfunk <- modarchive_download(41529)
  search_results <- modarchive_search_mod("*_-_intro.mod",
                                          size = "0-99",
                                          format = "MOD")
})

}
