% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmpt-package.R
\docType{package}
\name{openmpt-package}
\alias{openmpt}
\alias{openmpt-package}
\title{openmpt: Open ModPlug Tracker port for R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

}
\keyword{internal}
