% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{openmpt_info}
\alias{openmpt_info}
\title{Retrieve information about the OpenMPT library}
\usage{
openmpt_info(key = "library_version", ...)
}
\arguments{
\item{key}{A key \code{character} string indicating which information to retrieve.
can be \code{"library_version"}, \code{"library_features"} and many others. See
\href{https://lib.openmpt.org/doc/namespaceopenmpt_1_1string.html}{API documentation})
for all possible values.}

\item{...}{Ignored}
}
\value{
Returns a \code{character} string with the requested information.
}
\description{
A wrapper for the \code{get} function in libopenmpt (see
\href{https://lib.openmpt.org/doc/namespaceopenmpt_1_1string.html}{API documentation})
}
\examples{
openmpt_info("library_version")
openmpt_info("library_features")
openmpt_info("url")
}
