% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\title{Get a specific ModPlug pattern table}
\usage{
pattern(mod, pattern = 0L, width = 0L, pad = TRUE, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{pattern}{The pattern index (starting at \code{0}) of the pattern to get.}

\item{width}{The maximum number of characters the string should contain. \code{0} means no limit.}

\item{pad}{If \code{TRUE}, the string will be resized to the exact length provided in the width parameter.}

\item{...}{Ignored}
}
\value{
A \code{matrix} of pattern cells formatted as \code{character} strings. Each column represents
an audio channel.
}
\description{
Collects a specific pattern table from a tracker module and presents it as
a matrix of formatted \code{character} strings.
}
\examples{
mod <- demo_mod()
pattern(mod)
}
