% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch-tempo.R
\name{pitch-tempo}
\alias{pitch-tempo}
\alias{pitch_factor}
\alias{pitch_factor<-}
\alias{tempo_factor}
\alias{tempo_factor<-}
\title{Control the pitch and tempo of a module}
\usage{
pitch_factor(mod, ...)

pitch_factor(mod, ...) <- value

tempo_factor(mod, ...)

tempo_factor(mod, ...) <- value
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{...}{Ignored}

\item{value}{Replacement value. A \code{numeric} factor with which
to adjust the tempo or pitch of a module}
}
\value{
Returns current factor, or the updated object in case
of an assign operation (\verb{<-}).
}
\description{
Functions to control the pitch and tempo of a module.
}
\examples{
mod <- demo_mod()

## Increase the module pitch with a factor 2
pitch_factor(mod) <- 2
pitch_factor(mod)

## Increase the module tempo with a factor 2
tempo_factor(mod) <- 2
tempo_factor(mod)
}
