% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play a ModPlug Tracker module}
\usage{
play(mod, sample_rate = 44100L, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{sample_rate}{Output sample rate when playing the module.}

\item{...}{Ignored}
}
\value{
\code{NULL}
}
\description{
TODO
}
\examples{
\dontrun{
mod <- read_mod("https://api.modarchive.org/downloads.php?moduleid=41529#elektric_funk.mod")
play(mod)
}
}
