% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.openmpt}
\alias{print.openmpt}
\alias{format.openmpt}
\title{Implementation of basic S3 generics}
\usage{
\method{print}{openmpt}(x, ...)

\method{format}{openmpt}(x, ...)
}
\arguments{
\item{x}{Object to apply the method to.}

\item{...}{Ignored.}
}
\value{
In case of \code{print} and \code{format} a formatted string
with basic information about the module is returned.
}
\description{
Implementation of basic S3 generics such as \code{\link[=print]{print()}}.
}
