% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_duration_seconds}
\alias{get_duration_seconds}
\title{Get ModPlug Tracker module duration}
\usage{
get_duration_seconds(mod, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{...}{Ignored}
}
\value{
Returns a \code{numeric} value indicating the song duration in seconds.
}
\description{
Get the duration of the song from a \code{openmpt} class module object in seconds.
}
\examples{
mod <- demo_mod()
get_duration_seconds(mod)
}
