% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmpt-package.R
\docType{package}
\name{openmpt-package}
\alias{openmpt}
\alias{openmpt-package}
\title{openmpt: Open 'ModPlug' Tracker Port}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tracker music uses audio samples and pattern tables to organise musical compositions. Such music is stored in module files. This package reads, renders and plays module files using the 'libopenmpt' library \url{https://lib.openmpt.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/openmpt/}
  \item \url{https://github.com/pepijn-devries/openmpt}
  \item \url{https://lib.openmpt.org/}
  \item Report bugs at \url{https://github.com/pepijn-devries/openmpt/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Other contributors:
\itemize{
  \item Jeroen Ooms \email{jeroenooms@gmail.com} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID}) [contributor]
  \item Dirk Eddelbuettel (\href{https://orcid.org/0000-0001-6419-907X}{ORCID}) [contributor]
  \item Jester (Copyright holder of cyberrid.mod) [copyright holder, data contributor]
  \item OpenMPT Project Developers and Contributors [copyright holder]
}

}
\keyword{internal}
