% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\title{Get a specific ModPlug pattern table}
\usage{
pattern(mod, pattern = 0L, width = 0L, pad = TRUE, ...)
}
\arguments{
\item{mod}{A tracker module object of class \code{openmpt}.}

\item{pattern}{TODO}

\item{width}{TODO}

\item{pad}{TODO}

\item{...}{Ignored}
}
\value{
A \code{matrix} of pattern cells formatted as \code{character} strings. Each column represents
an audio channel.
}
\description{
TODO
}
\examples{
# TODO
}
